feat(storybook-addon): move util `correctImportPath` in common utils package
