+++ packages/utilities/src/index.ts
@@ -2,4 +2,5 @@ export * from './types';
 export * from './utils/common';
 export * from './utils/isEmpty';
 export * from './utils/importRemote';
+export * from './utils/correctImportPath';
 export * from './Logger';
